server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
pymongo==4.1.1
PyJWT==2.3.0
bcrypt==4.1.2
cryptography==41.0.4
werkzeug==2.0.3
python-multipart==0.0.6 